# services/body/server/openapi.yaml
openapi: 3.0.3
info:
  description: "API for delivering wellness content related to Body (e.g., Yoga videos,\
    \ exercise routines)."
  title: Body Service API
  version: 1.0.0
servers:
- url: http://localhost:8000
paths:
  /yoga/videos:
    get:
      description: Returns a list of available yoga videos with metadata.
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Video'
                type: array
          description: List of yoga videos
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Server error
      summary: Get available yoga videos
    post:
      description: Adds a new yoga video entry with metadata.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoUploadRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Video'
          description: Video uploaded successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid input
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Server error
      summary: Upload a new yoga video
components:
  schemas:
    Video:
      example:
        duration: 15:32
        difficulty: Beginner
        id: abc123
        title: Morning Flow Yoga
        url: https://cdn.wellnesshub.ai/videos/morning-flow.mp4
        tags:
        - flexibility
        - balance
      properties:
        id:
          example: abc123
          title: id
          type: string
        title:
          example: Morning Flow Yoga
          title: title
          type: string
        duration:
          example: 15:32
          title: duration
          type: string
        url:
          example: https://cdn.wellnesshub.ai/videos/morning-flow.mp4
          format: uri
          title: url
          type: string
        difficulty:
          enum:
          - Beginner
          - Intermediate
          - Advanced
          example: Beginner
          title: difficulty
          type: string
        tags:
          example:
          - flexibility
          - balance
          items:
            type: string
          title: tags
          type: array
      title: Video
      type: object
    VideoUploadRequest:
      example:
        duration: 20:00
        difficulty: Beginner
        title: Evening Relaxation Yoga
        url: https://cdn.wellnesshub.ai/videos/evening-relaxation.mp4
        tags:
        - tags
        - tags
      properties:
        title:
          example: Evening Relaxation Yoga
          title: title
          type: string
        url:
          example: https://cdn.wellnesshub.ai/videos/evening-relaxation.mp4
          format: uri
          title: url
          type: string
        duration:
          example: 20:00
          title: duration
          type: string
        difficulty:
          enum:
          - Beginner
          - Intermediate
          - Advanced
          title: difficulty
          type: string
        tags:
          items:
            type: string
          title: tags
          type: array
      required:
      - title
      - url
      title: VideoUploadRequest
      type: object
    Error:
      example:
        code: 500
        message: Internal Server Error
      properties:
        code:
          example: 500
          title: code
          type: integer
        message:
          example: Internal Server Error
          title: message
          type: string
      title: Error
      type: object