# services/body/server/src/src/openapi_server/models/video_upload_request.py
# coding: utf-8

"""
    Body Service API

    API for delivering wellness content related to Body (e.g., Yoga videos, exercise routines).

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations

import json
import pprint
import re  # noqa: F401
from typing import Any, ClassVar, Dict, List, Optional

from pydantic import BaseModel, ConfigDict, StrictStr, field_validator

try:
    from typing import Self
except ImportError:
    from typing_extensions import Self


class VideoUploadRequest(BaseModel):
    """
    VideoUploadRequest
    """  # noqa: E501

    title: StrictStr
    url: StrictStr
    duration: Optional[StrictStr] = None
    difficulty: Optional[StrictStr] = None
    tags: Optional[List[StrictStr]] = None
    __properties: ClassVar[List[str]] = [
        "title",
        "url",
        "duration",
        "difficulty",
        "tags",
    ]

    @field_validator("difficulty")
    def difficulty_validate_enum(cls, value):
        """Validates the enum"""
        if value is None:
            return value

        if value not in (
            "Beginner",
            "Intermediate",
            "Advanced",
        ):
            raise ValueError(
                "must be one of enum values ('Beginner', 'Intermediate', 'Advanced')"
            )
        return value

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }

    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of VideoUploadRequest from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={},
            exclude_none=True,
        )
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of VideoUploadRequest from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate(
            {
                "title": obj.get("title"),
                "url": obj.get("url"),
                "duration": obj.get("duration"),
                "difficulty": obj.get("difficulty"),
                "tags": obj.get("tags"),
            }
        )
        return _obj